refactor(gdn): split ssm_gdn_b at the wy4 seam; allow-list the one file with no seam
The GDN batched-verify stack (#837/#838/#844/#845, four PRs certified as one
group) pushed two files over the 500-LoC cap, which is a required context.

SPLIT: ssm_gdn_b.rs 391 -> 569. `gdn_decode_wy4`, `gdn_decode_wyn` and the
K=5..16 table form move verbatim into ssm_gdn_wyn.rs, leaving 205 + 399. Split
at wy4 rather than wy2 because a wy2 split left the sibling at 536 -- still over.
`ops.rs` declares and re-exports both halves, so no call site changes.

Verified exact, not asserted: 10 functions in, 3 + 7 out, zero bodies differing.

ALLOW-LISTED: qwen3_ssm/init.rs 481 -> 511. This one has no seam -- the file is
a single function (`impl Qwen3SsmLayer` at line 7, `pub fn new(` running to 504).
Splitting a 496-line constructor means restructuring it, which is a design call
for the SSM owner rather than a mechanical move by whoever tripped the cap. The
allow-list's own rule requires a rationale and a tracking issue; both are there
(#872), and the entry says why this file was treated differently from the two
that were split.
