+++ .github/workflows/file-size-cap.yml
@@ -56,6 +56,15 @@ jobs:
           # split. Adding to this list requires a rationale comment AND
           # a tracking issue.
           allow_list=(
+            # 2026-09-04 GDN batched-verify stack (#837/#838/#844/#845): 481 -> 511
+            # LoC from the K=5..16 pointer-table init. NOT split because the file is
+            # ONE function -- `impl Qwen3SsmLayer` opens at line 7 and `pub fn new(`
+            # runs to 504, so there is no piecewise seam. The two other files this
+            # stack pushed over the cap WERE split (verify_a.rs 638 -> 417+254,
+            # ssm_gdn_b.rs 569 -> 205+399, bodies verified byte-identical).
+            # Splitting a 496-line constructor is a design decision for the SSM
+            # owner, not a mechanical move. Tracked: #872
+            "crates/spark-model/src/layers/qwen3_ssm/init.rs"
             # 2026-08-15 concurrency cycle (#525): grew past 500 in this stack's
             # instrumentation work. Reduce when feasible.
             "crates/atlas-plugin/src/benchmarks/concurrency.rs"
