+++ crates/spark-model/src/layers/ops.rs
@@ -145,6 +145,8 @@ mod ssm_gdn_b;
 mod ssm_gdn_batched;
 #[path = "ops/ssm_gdn_snap.rs"]
 mod ssm_gdn_snap;
+#[path = "ops/ssm_gdn_wyn.rs"]
+mod ssm_gdn_wyn;
 #[path = "ops/ssm_mamba.rs"]
 mod ssm_mamba;
 #[path = "ops/ssm_preproc.rs"]
@@ -211,6 +213,7 @@ pub use ssm_gdn_a::*;
 pub use ssm_gdn_a2::*;
 pub use ssm_gdn_a3::*;
 pub use ssm_gdn_b::*;
+pub use ssm_gdn_wyn::*;
 pub use ssm_gdn_batched::*;
 pub use ssm_gdn_snap::*;
 pub use ssm_mamba::*;
